chore(tsc): typesafe test render utils
